Updated docker run command and added new variables
Added TZ environment variable to the Docker Image to allow users to change the timezone of the container to ensure logs use the preferred date and time. Default is UTC.

Added ENVIRONMENT environment variable to the Docker Image which sends radsecproxy logs to stdout which allows the user to view Docker logs from ```docker logs <container>```.

Removed unnecessary `-it` flag when starting a container in the background.